requirements.txt
Flask
Flask-Cors
Pillow
gunicorn